+++ src/main.rs
@@ -13,7 +13,7 @@ fn main() {
     pretty_env_logger::init();
 
     let rt = Runtime::new().unwrap();
-    rt.block_on(persistence::rdbms::setup()).unwrap();
+    rt.block_on(persistence::rdbms_init::setup()).unwrap();
 
     // run the api to load the records for inspection
     match rt.block_on(suma::sumaclient::run_api()) {
