+++ src/persistence/mod.rs
@@ -1,1 +1,2 @@
 pub mod rdbms;
+pub mod rdbms_init;
