+++ src/persistence/rdbms.rs
@@ -1,14 +1,13 @@
 use futures::TryStreamExt;
-use sqlx::{Row, Connection, ConnectOptions, Result};
+use sqlx::{Row, Result};
 use rusqlite::{Connection as rusqliteConnection, params, OpenFlags};
 use once_cell::sync::OnceCell;
-use sqlx::sqlite::{SqliteConnectOptions, SqlitePool};
+use sqlx::sqlite::{SqlitePool};
 use tokio::runtime::Runtime;
 use std::collections::HashMap;
-use std::str::FromStr;
 
 // Constants
-const DB_NAME: &str = "suse_errata.sqlite";
+pub const DB_NAME: &str = "suse_errata.sqlite";
 const POOL: OnceCell<SqlitePool> = OnceCell::new();
 
 // ENUM
@@ -70,67 +69,12 @@ fn init_pool() -> SqlitePool {
 }
 
 // current path
-fn db_path() -> String {
+pub fn db_path() -> String {
     let dir = std::env::current_exe().unwrap().parent().expect("Executable must be in some directory").to_owned();
     let path = dir.join("db");
 
     // return path string
     path.to_str().unwrap().to_owned()
-}
-
-pub async fn setup() -> Result<()> {
-    // create db path if it not exists
-    if !std::path::Path::new(db_path().as_str()).exists() {
-        match std::fs::create_dir(db_path().as_str()) {
-            Err(e) => log::error!("Unable to create directory: {}", e),
-            Ok(()) => log::info!("Directory created"),
-        };
-    }
-
-    let setup_stmt = r#"
-        begin;
-            CREATE TABLE IF NOT EXISTS errata (
-                suse_id INTEGER UNIQUE,
-                date REAL NOT NULL,
-                update_date REAL NOT NULL,
-                advisory_type TEXT,
-                advisory_name TEXT NOT NULL UNIQUE,
-                advisory_synopsis TEXT
-            );
-            CREATE TABLE IF NOT EXISTS cve (
-                id TEXT PRIMARY KEY,
-                score REAL
-            );
-            CREATE TABLE IF NOT EXISTS errata_cve (
-                id INTEGER PRIMARY KEY,
-                cve_id TEXT NOT NULL REFERENCES cve(id),
-                errata_id INTEGER NOT NULL REFERENCES ERRATA(suse_id)
-            );
-            CREATE TABLE IF NOT EXISTS cve_systems (
-                id INTEGER PRIMARY KEY,
-                cve_id TEXT NOT NULL REFERENCES cve(id),
-                affected_systems_cnt INTEGER,
-                last_checked_date REAL
-            );
-        COMMIT;
-    "#;
-
-    let full_path = format!("{}/{}", db_path(), DB_NAME);
-
-    println!("full_path is: {}", &full_path);
-
-    let mut conn = SqliteConnectOptions::from_str(&full_path)?
-        .create_if_missing(true)
-        .connect().await?;
-    
-    _= match sqlx::query(setup_stmt).execute(&mut conn).await {
-        Ok(_)   => Ok(()),
-        Err(e) => Err(e),
-    };
-
-    conn.close();
-
-    Ok(())
 }
 
 pub async fn batch_store(
