+++ src/tests.rs
@@ -1,13 +1,19 @@
 #[cfg(test)]
 use cve_audit_types::*;
-use crate::persistence::rdbms;
+use crate::persistence::*;
 use std::collections::HashMap;
 
 #[test]
 fn sql_demo_data() {
     let rt = tokio::runtime::Runtime::new().unwrap();
 
-    let _ok = rdbms::setup();
+    let setup = async {rdbms_init::setup().await};
+    let setup_res = match rt.block_on(setup) {
+        Ok(_) => true,
+        Err(_) => false,
+    };
+    assert!(setup_res);
+
     // add some code to add demo data
     let mut errata:Vec<ErrataExtended> = Vec::new();
     // first some errata
